Created user#show test
