+++ spec/support/authentication_helper.rb
@@ -1,0 +1,9 @@
+module AuthenticationHelper
+  def sign_in_as(user)
+    visit root_path
+    click_link "Sign In"
+    fill_in 'Email', with: user.email
+    fill_in 'Password', with: user.password
+    click_button 'Log in'
+  end
+end
